types/{segment,checker_exec}: use u64 for segment/exec ids

+++ reldrv/src/types/checker_exec.rs
@@ -16,7 +16,7 @@ use crate::{
     process::{state::Stopped, Process},
 };
 
-pub type CheckerExecutionId = u32;
+pub type CheckerExecutionId = u64;
 
 #[derive(Debug)]
 pub struct CheckerExecution {
